feat(migrations): add migration to update cigars filler and binder columns
The migration file `2023_09_09_131004_update_cigars_filler_binder.php` has been added. This migration updates the `filler` and `binder` columns in the `cigars` table to be nullable. It also adds foreign key constraints to these columns.

The migration also updates the existing data in the `cigars` table. It retrieves all cigars and for each cigar, it finds the corresponding wrapper by name and sets the `binder` and `filler` columns to the wrapper's ID. If a wrapper with the given name is not found, the columns are set to null.

In the `down` method of the migration, the changes are reversed. The `filler` and `binder` columns are set back to being non-nullable and their values are updated based on the wrapper's name.

fix(Create.vue): remove unused code and update input fields

In the `Create.vue` file, the unused code for `typeRules`, `originRules`, `strengthRules`, and `flavorProfileRules` has been removed.

The input fields for `binder` and `filler` have been updated to use `v-autocomplete` instead of `v-text-field`. The `wrapperOptions` are passed as items to the `v-autocomplete` components. The `binder` and `filler` variables are bound to the selected values in the autocomplete fields. The `binderRules` and `fillerRules` are applied to validate the input.